1.4-docker-compose/crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
python-dotenv
gunicorn
requests
